Add AppointmentIndicatorService import to services initialization
This commit updates the __init__.py file in the services module to include the AppointmentIndicatorService. The addition allows for the AppointmentIndicatorService to be accessible when the services package is imported, enhancing the functionality of the module. With this change, other components of the application can utilize the appointment indicator service without needing to import it directly from its module.