Add AppointmentStateTimelineService import to core services
This commit updates the core services module by importing the newly created AppointmentStateTimelineService. This addition enables the use of the AppointmentStateTimelineService functionality within the core services package, enhancing the overall capabilities of the module. The import statement has been added to the __init__.py file to ensure that the service is accessible at the package level.