Add AppointmentReviewService import to core services
This commit introduces the import of AppointmentReviewService in the __init__.py file of the core services module. By adding this import, we ensure that the AppointmentReviewService is accessible as part of the core services package. This change is intended to facilitate better management of appointment reviews within the application, enhancing the overall functionality and organization of the service layer.